requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==24.3.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
dataclasses-json==0.6.7
Deprecated==1.2.15
distro==1.9.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
limits==4.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.25.1
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.59.8
orjson==3.10.15
packaging==24.2
passlib==1.7.4
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.10.1
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
shellingham==1.5.4
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.37.2
tenacity==9.0.0
tqdm==4.67.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
urllib3==2.3.0
uvicorn==0.30.1
watchfiles==0.22.0
websockets==12.0
wrapt==1.17.2
yarl==1.18.3
